Refactor statistics
Now also executable has `statistics` feature that needs to be enabled compile-time to collect
information about largest/smallest numbers that appear in the games. It is useful to choose correct
type for the representation, currently we're on safe/conservative side at the cost of memory. Would
be best to reduce `Nus` to 12 bytes, to match `GamePtr` + enum discriminant size