Separate transposition table by no. of empty places
This enables exhaustive search to be done in order from least to most empty places. E.g. Domineering
position with `n` empty spaces has moves to positions with `n-2` empty spaces, so we can prune
everything below it out of memory. The same holds for snort, if we have `n` empty/tinted vertices,
we're only interested in positions with `n-1` empty/tinted vertices, as they follow directly in the
game tree.

This is a performace hit for a tradeof of actually working with under 64 GB of RAM.

This needs in-code more docs.