Add `TotalWrapper`
It is useful when we need to stick some game type in a collection. Then we can make the collection
operate on `TotalWrapper`s which do not follow game algebra that we would expect. That way we can
not implement common ordering and hashing traits on game types that have partial ordering at best

We have `coerce` at home